test(solid-query): use vitest typecheck correctly with *.test-d.ts
